[기능 구현] 임수영 - 강사 챕터 삭제
