+++ src/main/resources/templates/course/deletechapter.html
@@ -1,0 +1,132 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+  <meta charset="UTF-8">
+  <title>챕터 삭제 확인</title>
+  <style>
+    body {
+        font-family: Arial, sans-serif;
+        max-width: 600px;
+        margin: 50px auto;
+        padding: 20px;
+    }
+
+    .container {
+        border: 2px solid #ddd;
+        border-radius: 10px;
+        padding: 30px;
+        background-color: #f9f9f9;
+    }
+
+    h2 {
+        color: #d32f2f;
+        text-align: center;
+    }
+
+    .chapter-info {
+        background-color: white;
+        padding: 20px;
+        border-radius: 5px;
+        margin: 20px 0;
+    }
+
+    .info-row {
+        margin: 10px 0;
+        padding: 10px 0;
+        border-bottom: 1px solid #eee;
+    }
+
+    .info-row:last-child {
+        border-bottom: none;
+    }
+
+    .label {
+        font-weight: bold;
+        color: #555;
+        display: inline-block;
+        width: 120px;
+    }
+
+    .warning {
+        background-color: #fff3cd;
+        border: 1px solid #ffc107;
+        padding: 15px;
+        border-radius: 5px;
+        margin: 20px 0;
+        color: #856404;
+    }
+
+    .button-group {
+        text-align: center;
+        margin-top: 30px;
+    }
+
+    button {
+        padding: 12px 30px;
+        margin: 0 10px;
+        font-size: 16px;
+        border: none;
+        border-radius: 5px;
+        cursor: pointer;
+    }
+
+    .btn-delete {
+        background-color: #d32f2f;
+        color: white;
+    }
+
+    .btn-delete:hover {
+        background-color: #b71c1c;
+    }
+
+    .btn-cancel {
+        background-color: #757575;
+        color: white;
+    }
+
+    .btn-cancel:hover {
+        background-color: #616161;
+    }
+  </style>
+</head>
+<body>
+
+<div class="container">
+  <h2>챕터 삭제 확인</h2>
+
+  <div class="chapter-info">
+    <div class="info-row">
+      <span class="label">강의명:</span>
+      <span th:text="${course.courseTitle}"></span>
+    </div>
+    <div class="info-row">
+      <span class="label">챕터명:</span>
+      <span th:text="${chapter.chapTitle}"></span>
+    </div>
+    <div class="info-row">
+      <span class="label">챕터 설명:</span>
+      <span th:text="${chapter.chapDesc}"></span>
+    </div>
+  </div>
+
+  <div class="warning">
+    <strong>⚠ 경고:</strong> 이 작업은 되돌릴 수 없습니다.<br>
+    해당 챕터가 영구적으로 삭제됩니다.
+  </div>
+
+  <div class="button-group">
+    <form th:action="@{/teacher/course/deletechapter}" method="post" style="display:inline;">
+      <input type="hidden" name="chapNo" th:value="${chapter.chapNo}">
+      <input type="hidden" name="courseId" th:value="${course.courseId}">
+      <button type="submit" class="btn-delete">삭제하기</button>
+    </form>
+
+    <form action="/teacher/course/chapter" method="post" style="display:inline;">
+      <input type="hidden" name="courseId" th:value="${course.courseId}">
+      <button type="submit" class="btn-cancel">취소</button>
+    </form>
+  </div>
+</div>
+
+</body>
+</html>
